Lost count on the changes. Added stuff, building.

+++ Calc/Impl/MonthlyLedger.cs
@@ -9,11 +9,11 @@ namespace PetShopLib.Impl
 {
     internal class MonthlyLedger : IMonthlyLedger
     {
-        public int Year { get; set; }
+        public short Year { get; set; }
         public short Month { get; set; }
         public decimal Income { get; set; }
         public decimal Expenses { get; set; }
-        public decimal Total { get; set; }
+        public decimal Total { get; set; }        
 
         public MonthlyLedger()
         {
